Add mocks for`User` and `Case` for tests
- The behaviour of sequelize-mock is such that it returns the value
- Edit the `user` route test cases to anticipate the expected changes
