+++ api/src/tests/fixtures/users.json
@@ -1,10 +1,12 @@
 [
   {
+    "id": 1,
     "name": "admin",
     "email": "admin@opengov.com",
     "agency": "MOE"
   },
   {
+    "id": 2,
     "name": "admin2",
     "email": "admin2@opengov.com",
     "agency": "HDB"
