+++ api/src/tests/helpers/initialise-mocks.ts
@@ -1,0 +1,18 @@
+import UserMock from '../mocks/User'
+import CaseMock from '../mocks/Cases'
+
+export const initialiseMocks = () => {
+  jest.mock('../../models/User', () => {
+    return {
+      _esModule: true,
+      default: UserMock,
+    }
+  })
+
+  jest.mock('../../models/Case', () => {
+    return {
+      _esModule: true,
+      default: CaseMock,
+    }
+  })
+}
