+++ api/src/tests/mocks/Cases.ts
@@ -1,0 +1,7 @@
+let SequelizeMock = require('sequelize-mock')
+let DBConnectionMock = new SequelizeMock()
+let cases = require('../fixtures/cases.json')
+
+let CaseMock = DBConnectionMock.define('cases', cases[0])
+
+export default CaseMock
